requirements.txt
streamlit>=1.31
pandas>=2.0
numpy>=1.24
plotly>=5.18
python-dotenv>=1.0

hopsworks[python]>=4.6,<4.7
